Split potoken.py into multiple files
