+++ docker/scripts/startup.sh
@@ -10,5 +10,5 @@ sleep 2
 
 echo "[INFO] launching chromium instance"
 
-# Run python script on display 99
-DISPLAY=:99 python index.py
+# Run python script on display 0
+DISPLAY=:99 python potoken-generator.py --oneshot
