+++ potoken-generator.py
@@ -1,0 +1,4 @@
+import potoken_generator.main
+
+if __name__ == '__main__':
+    potoken_generator.main.main()
